Initialze eLxr Pro package
Establish the codebase for the 'elxr-pro' application and implement
several commands, with a focus on associating and disassociating the
target system from an organization's entitlements.

The first version is just to generate the deb package. After it was
installed into the target, the subcommands can be executed without
the real actions.

At default, the binary command elxr-pro will be installed at the path
/usr/bin with the installation 'dpkg -i elxr-pro_.*.deb'.

Support the following sub-commands(join/leave/test/config):
 $ which elxr-pro
 /usr/bin/elxr-pro
 $ elxr-pro
 usage: elxr-pro [-h] [--debug] [--version] <command> ...

 Quick start commands:

   join             attach this machine to an eLxr Pro subscription

 Other commands:

   config           manage eLxr Pro configuration on this machine
   leave            remove this machine from an eLxr Pro subscription
   test             validate the connection to the API server

 Flags:

   -h, --help       Displays help on elxr-pro and command line options
   --debug          show all debug log messages to console
   --version        show version of elxr-pro

 Use elxr-pro <command> --help for more information about a command.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>