Add the gpg signature example file
To secure access to the elxr pro package repository, elxr-pro
client provide the repository signature gpg file.
In the future we will add the real gpg files into the keyring
directory.
Here we can check if the gpg file will be installed into the
correct path /usr/share/keyrings in the target.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>