Add the unit-tests and fix few issues
To support the unit tests feature, now we support the pytest to run
all of the unit tests and to generate the coverage report.

The report will be at /usr/share/doc/elxr-pro/coverage/index.html
when the elxr-pro deb package is installed.

The latest unit test coverage rate is about 90% in the report.
Please refer to the detailed information after installing the package
or unpacking the deb package with dpkg command.

At the same time fix few small issues and typos.

TestPlan:
As we have no the available pipeline config file to run the full path
to build deb package in the elxr-pro project, so I do the validation
locally with the following steps:
 $ cd $project_dir
 $ apt build-dep .
 $ dpkg-buildpackage

Check the deb package can be generated successfully.
Check the unit test coverage report has been generated successfully
with the following command:
 $ dpkg -X elxr-pro_1.0.0~elxr12_all.deb $unpack_dir
 $ tree $unpack_dir/usr/share/doc/elxr-pro/coverage/

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>