Add the public gpg key file for elxr pro
As a temporary solution, we install the elxr public gpg key into
the deb package as the pro entitlements public key.
We can use the accurate key files instead in the future when it's
ready.
After the deb package was installed into the target, this key will
be located at /usr/share/keyrings directory.

If the key files are going to be prepared in the post of 2506, we
can import it by manual in the target.

Meanwhile drop the original fake gpg file.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>