Drop the useless argument enable_services
This item has been removed from the fields, but it's still
left in the __init__ function as a argument.
So that it's acquired with exception when we use the option
--attach-config for join command.
Now we cleanup it from the __init__ function to avoid the
incorrect requirement.

TestPlan:
PASS: dpkg-buildpackage
PASS: install in the target: $ dpkg -i elxr-pro*.deb
PASS: $cat auth-config.txt
      token: abcdefghijklmnopqrstuvwxyz1234567890
PASS: sudo elxr-pro join --attach-config=./auth-config.txt
PASS: sudo elxr-pro leave

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>