debian: update salsa-ci.yml file to use the elxrpro salsa-ci pipeline
The current salsa-ci.yml file makes use of the salsa-ci pipeline from
the elxr project. Although these pipelines are similar and most jobs
are the same, it will fail when we attempt to publish into Aptly since
some of the variables and script are specific to elxr or elxrpro.

Additionally the current implementation of the salsa-ci.yml file is
faulty as we should not be directly including salsa-ci.yml and
pipeline-jobs.yml from the pipeline project directly but rather
including recipes/elxrpro.yml This allows for changes to the recipe in
the pipeline project to easily propagate to all projects without
having to edit every project themselves.

We also can't use the URL method used in the elxr project, since the
pipeline project is private, we must use the project,file,ref method
instead. This will ensure we can access the recipe file in the
pipeline project.

Lastly, since the variables being set are already in the elxrpro.yml
recipe we actually don't need to define again, so drop them here.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>