Revert "Disable to publish elxr-pro deb to aptly"
This reverts commit 64adf225a406a0f98e2cf9b1700018f95146780f.

As we only need to download the elxr-pro cli from the Gallery in
the first time, after the connection to API Server and get the
eLxr Pro entitlements, after that we can upgrade the client from
Pro resource with the command 'apt install elxr-pro'.
So just revert the original commit to recover the uploading to
aptly repo.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>